Change some types to uint64_t, restore borrowedsort.

+++ include/utilities/Constants.h
@@ -10,8 +10,8 @@ namespace Constants {
 
     const uint16_t N_PARTITIONS = 1024;
 
-    const uint16_t KEY_SIZE   = 10;
-    const uint16_t TUPLE_SIZE = 100;
+    const uint64_t KEY_SIZE   = 10;
+    const uint64_t TUPLE_SIZE = 100;
     const uint16_t CHUNK_SIZE = 4000; // If CHUNK_SIZE is 0, the whole thread segment will be read
 }
